requirements.txt
pywebview>=4.0.0
ollama>=0.1.0
